# rotator-server/static/openapi.yaml
openapi: 3.0.3
info:
  title: Antenna Rotator Server
  version: "2.0.0"
  description: |
    HTTP API for controlling a serial-attached antenna rotator (Prosistel or
    Yaesu GS-232 protocol). Use this Swagger UI to read or set the antenna
    heading, stop in-progress rotation, and inspect the available serial
    ports.

    The canonical paths live under `/api/v1`; the unprefixed paths remain
    available as legacy aliases.

    Start the server with `SIMULATION=true` to run against an in-memory
    rotator (no hardware required) — useful for testing and demos.
servers:
  - url: /api/v1
    description: Canonical API prefix
  - url: /
    description: Legacy unprefixed paths
tags:
  - name: rotator
    description: Rotator control endpoints
  - name: system
    description: Health and diagnostics
paths:
  /set-heading:
    post:
      tags: [rotator]
      summary: Set the antenna heading
      description: |
        Rotate the antenna to the supplied azimuth (0–359 degrees). Pass the
        heading either as the `heading` query parameter or as a JSON body.
      parameters:
        - in: query
          name: heading
          required: false
          description: Target azimuth in degrees (0–359). Takes precedence over the body.
          schema:
            type: integer
            minimum: 0
            maximum: 359
            example: 180
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              required: [heading]
              properties:
                heading:
                  type: integer
                  minimum: 0
                  maximum: 359
                  example: 180
      responses:
        "200":
          description: Heading accepted and command dispatched.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HeadingResponse"
              example:
                heading: 180
                status: set
        "400":
          $ref: "#/components/responses/BadRequest"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "502":
          $ref: "#/components/responses/RotatorUnavailable"
  /get-heading:
    get:
      tags: [rotator]
      summary: Read the current antenna heading
      responses:
        "200":
          description: Current azimuth in degrees.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HeadingResponse"
              example:
                heading: 180
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "502":
          $ref: "#/components/responses/RotatorUnavailable"
  /stop:
    post:
      tags: [rotator]
      summary: Stop any in-progress rotation
      responses:
        "200":
          description: Stop command dispatched.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StopResponse"
              example:
                status: stopped
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "502":
          $ref: "#/components/responses/RotatorUnavailable"
  /offset:
    get:
      tags: [rotator]
      summary: Read the current calibration offset
      description: |
        Real-world heading = raw rotator heading + offset (mod 360). Useful
        for correcting a mechanically misaligned mounting so that requested
        headings point where you actually expect.
      responses:
        "200":
          description: Current calibration offset in degrees.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OffsetResponse"
              example:
                offset: 15
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
    post:
      tags: [rotator]
      summary: Set the calibration offset
      description: |
        Applies to all subsequent set-heading/get-heading calls. Pass the
        offset either as the `offset` query parameter or as a JSON body. Any
        integer is accepted and normalised modulo 360; takes effect
        immediately, with no hardware command sent.
      parameters:
        - in: query
          name: offset
          required: false
          description: Calibration offset in degrees. Takes precedence over the body.
          schema:
            type: integer
            example: 15
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              required: [offset]
              properties:
                offset:
                  type: integer
                  example: 15
      responses:
        "200":
          description: Offset applied.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OffsetResponse"
              example:
                offset: 15
                status: set
        "400":
          $ref: "#/components/responses/BadRequest"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /list-ports:
    get:
      tags: [system]
      summary: List available serial ports
      description: Returns every serial port the host currently exposes.
      responses:
        "200":
          description: List of detected serial ports.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortsResponse"
              example:
                ports: ["COM3", "COM4"]
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/ServerError"
  /healthz:
    get:
      tags: [system]
      summary: Liveness probe
      description: |
        Always returns 200 while the process is alive. Inspect `mode` and
        `connected` to distinguish a healthy hardware link from simulation
        mode or a dropped serial connection.
      responses:
        "200":
          description: Server is running.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              example:
                status: ok
                mode: hardware
                protocol: yaesu-gs232
                connected: true
                version: "1.2.0"
components:
  schemas:
    HeadingResponse:
      type: object
      required: [heading]
      properties:
        heading:
          type: integer
          minimum: 0
          maximum: 359
          description: Azimuth in degrees.
        status:
          type: string
          description: Action acknowledgement (only present on writes).
          enum: [set]
    StopResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [stopped]
    OffsetResponse:
      type: object
      required: [offset]
      properties:
        offset:
          type: integer
          description: Calibration offset in degrees (real-world heading = raw + offset, mod 360).
        status:
          type: string
          description: Action acknowledgement (only present on writes).
          enum: [set]
    PortsResponse:
      type: object
      required: [ports]
      properties:
        ports:
          type: array
          items:
            type: string
          description: Serial port identifiers (e.g. `COM3`, `/dev/ttyUSB0`).
    HealthResponse:
      type: object
      required: [status, mode, protocol, connected, version]
      properties:
        status:
          type: string
          enum: [ok]
        mode:
          type: string
          enum: [hardware, simulation]
          description: Whether a real serial rotator is being driven.
        protocol:
          type: string
          enum: [prosistel, yaesu-gs232, simulation]
          description: Wire protocol in use.
        connected:
          type: boolean
          description: Whether the serial port is currently open (always true in simulation mode).
        version:
          type: string
          description: Build version of the running server.
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
  responses:
    BadRequest:
      description: Invalid request (e.g. heading out of range or missing).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "invalid heading: 400 out of range, must be 0–359"
    MethodNotAllowed:
      description: HTTP method is not supported on this route.
      content:
        text/plain:
          schema:
            type: string
            example: Method Not Allowed
    ServerError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: could not list serial ports
    RotatorUnavailable:
      description: Serial communication with the rotator failed (device unplugged, timeout, etc.).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "failed to read heading: timeout: no response from rotator"